header only and can handle double or float
