+++ src/CMakeLists.txt
@@ -5,6 +5,6 @@ project (example)
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_BUILD_TYPE RelWithDebInfo)
 
-set(SOURCES example.cpp resample.cpp)
+set(SOURCES example.cpp)
 
 add_executable(${PROJECT_NAME} ${SOURCES})
